Changed the way we drag with a scroll view
By leveraging raw events from Listener instead of GestureDetector batteling in the GestureArena we are able to be less dependent on the Notification events to controll the dragging of the sheet and we can instead use vertical drag gestures directly to control the sheet.

There is also added parameters to control the default scrollphysics of underlaying scrollable views, this lets us disable scrolling in the ScrollPhysics implementation from userspace.